+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.7.0 (October 6, 2022). Tested on Artifactory 7.41.13 and Xray 3.57.6
+
+NEW FEATURE:
+
+* resource/xray_repository_config: add a new resource allowing to configure Xray report retention policies for the repositories. Issue [#77](https://github.com/jfrog/terraform-provider-xray/issues/77) PR [#81](https://github.com/jfrog/terraform-provider-xray/pull/81)
+* Add ability to disable license check. PR [#80](https://github.com/jfrog/terraform-provider-xray/pull/80)
+
 ## 1.6.0 (August 31, 2022). Tested on Artifactory 7.41.7 and Xray 3.55.2
 
 NEW FEATURE: